Add WorkerInfo.h header file
